requirements.txt
Flask==3.0.3
nltk==3.5
numpy==1.19.5
tensorflow==2.4.1
mysql-connector-python==8.0.26
Sastrawi==1.0.1
pandas==1.1.5
scikit-learn==0.23.2
scipy==1.10.1